Initial working observation import against sampling period
