+++ api/src/paths/project/{projectId}/survey/{surveyId}/observations/upload.ts
@@ -56,6 +56,7 @@ POST.apiDoc = {
         schema: {
           type: 'object',
           additionalProperties: false,
+          required: ['media'],
           properties: {
             media: {
               description: 'A survey observation submission file.',
