+++ app/src/features/surveys/observations/SurveyObservationPage.tsx
@@ -8,7 +8,7 @@ import { SurveyContext } from 'contexts/surveyContext';
 import { TaxonomyContextProvider } from 'contexts/taxonomyContext';
 import { useContext } from 'react';
 import ObservationsTableContainer from './observations-table/ObservationsTableContainer';
-import SamplingSiteList from './sampling-sites/SamplingSiteList';
+import SamplingSiteList from './sampling-sites/list/SamplingSiteList';
 import SurveyObservationHeader from './SurveyObservationHeader';
 
 export const SurveyObservationPage = () => {
